+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.5.2 / ####-##-##
 ==================
 
+  * Suites can now be reordered
   * Run/Stop option in context menu
   * Duplicate task option in context menu
   * Fixed delete suite bug